Code simplification + get overlapping region
Store the points as Points rather than individual primitives.  Also, there  was no reason to make the upper-right a calculated value when it can just be specified explicitly.

Added method to get the rectangle defining the overlapping region.  This is an intermediate step to getting just the points of intersection, but it makes for a useful public API method.